fix: refresh the customer token in middleware
