+++ examples/nextjs/src/middlewares/withCommerceLayer.ts
@@ -2,21 +2,28 @@ import {
   authenticate,
   getCoreApiBaseEndpoint,
   jwtVerify,
+  makeSalesChannel,
 } from "@commercelayer/js-auth"
 import { NextResponse } from 'next/server'
+import { createStorage } from "unstorage"
+import memoryDriver from "unstorage/drivers/memory"
+import { makeCustomerStorage } from "@/app/utils/customerStorage"
 import type { WithMiddleware } from "./types"
 
-export const withCommerceLayer: WithMiddleware = (next = () => NextResponse.next()) => {
-  return async (request, event) => {
+const clientId = process.env.NEXT_PUBLIC_CLIENT_ID as string
+const scope = process.env.NEXT_PUBLIC_SCOPE as string
+
+export const withCommerceLayer: WithMiddleware = (
+  _next = () => NextResponse.next(),
+) => {
+  return async (request, _event) => {
     if (request.nextUrl.pathname.endsWith("/middleware")) {
       const auth = await authenticate("client_credentials", {
-        clientId: process.env.NEXT_PUBLIC_CLIENT_ID as string,
-        scope: process.env.NEXT_PUBLIC_SCOPE as string,
+        clientId,
+        scope,
       })
 
       const decodedJWT = await jwtVerify(auth.accessToken)
-
-      let slug: string
 
       if (!("organization" in decodedJWT.payload)) {
         throw new Error('A "sales_channel" token is required.')
@@ -27,6 +34,31 @@ export const withCommerceLayer: WithMiddleware = (next = () => NextResponse.next
         orgSlug: decodedJWT.payload.organization.slug,
         baseEndpoint: getCoreApiBaseEndpoint(auth.accessToken),
       })
+    }
+
+    if (request.nextUrl.pathname.endsWith("/api-credentials/server")) {
+      /**
+       * NOTE: This is a workaround to make the `refresh_token` working properly.
+       * When custom access token expires, the `getAuthorization` method
+       * will automatically refresh the access token using the `refresh_token`.
+       * However, the customer token is stored in cookies and cookies
+       * can only be set in Server Actions or Route Handlers.
+       */
+      const salesChannel = makeSalesChannel(
+        {
+          clientId,
+          scope,
+          debug: true,
+        },
+        {
+          storage: createStorage({
+            driver: memoryDriver(),
+          }),
+          customerStorage: makeCustomerStorage(),
+        },
+      )
+
+      await salesChannel.getAuthorization()
     }
 
     return NextResponse.next()
